avance

+++ db/migrate/20250205141336_create_agrochemicals.rb
@@ -2,7 +2,7 @@ class CreateAgrochemicals < ActiveRecord::Migration[6.1]
   def change
     create_table :agrochemicals do |t|
       t.string  :nombre,      null: false
-      t.integer :cantidad, null: false
+      t.decimal :cantidad, null: false
       t.text    :ingrediente_activo, null: false
       t.text    :objetivo, null: false
       t.references :agrochemical_division, null: false, foreign_key: true
@@ -12,11 +12,6 @@ class CreateAgrochemicals < ActiveRecord::Migration[6.1]
       t.integer :reingreso, null: false
       t.text    :daño_a_abejorros, null: false
 
-      
-
-      #Generar una tabla de division
-      #Herbicidas, Acaricidas, Fungicidas, Insecticidas, Nematicidas, Rodenticidas, Coadyuvantes
-      
       t.timestamps
     end
   end
